Antiscroll not cleaning up correctly after itself
On line 43, Antiscroll set's the width and height of `.antiscroll-inner` element and on line 58 it uses those dimensions to determine if scrollbars needs added. Whenever you call `$(selector).antiscroll()` the plugin destroys the Antiscroll instance but doesn't reset the dimensions of `.antiscroll-inner` element.

This creates very subtle problems with resizing tables. 

The solution is to do proper cleanup. 

Note: this effects how Ember Table rebuilds the antiscroll because you don't need to call `.rebuild` you can just reinstantiate Antiscroll. I'll submit a PR for that as well.